refact(step2): clarify service and controller
