+++ src/main/java/step2/LottoMain.java
@@ -4,7 +4,7 @@ import step2.controller.LottoController;
 
 public class LottoMain {
     public static void main(String[] args) {
-        LottoController lottoController = new LottoController();
-        lottoController.runManualLotto();
+        LottoController lottoService = new LottoController();
+        lottoService.buyManualLotto();
     }
 }
